requirements.txt
alembic==1.6.0
amqp==5.0.6
attrs==21.2.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cowin==0.0.5
decorator==5.0.7
fake-useragent==0.1.11
fastapi==0.64.0
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-PyMongo==2.3.0
googlemaps==4.4.5
greenlet==1.0.0
h11==0.12.0
haversine==2.3.0
httpcore==0.13.3
httpx==0.18.1
idna==2.10
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mysql==0.0.2
mysqlclient==2.0.3
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pydantic==1.8.1
Pygments==2.9.0
pyIsEmail==1.3.2 
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.13
sqlalchemy_utils==0.37.2
starlette==0.13.6
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
